added db communication to dashboard
